first pass at splitting users from members
