+++ src/domain/ecoverse/ecoverse.module.ts
@@ -10,7 +10,6 @@ import { TagsetModule } from '@domain/tagset/tagset.module';
 import { ChallengeModule } from '@domain/challenge/challenge.module';
 import { UserModule } from '@domain/user/user.module';
 import { OrganisationModule } from '@domain/organisation/organisation.module';
-import { AccountModule } from '@utils/account/account.module';
 import { ApplicationModule } from '@domain/application/application.module';
 
 @Module({
@@ -25,7 +24,6 @@ import { ApplicationModule } from '@domain/application/application.module';
     UserModule,
     TypeOrmModule.forFeature([Ecoverse]),
     UserModule,
-    AccountModule,
     ApplicationModule,
   ],
   providers: [
