+++ src/domain/user/user.resolver.spec.ts
@@ -1,19 +1,25 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AppModule } from '@src/app.module';
-import { UserResolver } from './user.resolver';
+import { UserResolverQueries } from './user.resolver.queries';
+import { UserResolverMutations } from './user.resolver.mutations';
 
 describe('UserResolver', () => {
-  let resolver: UserResolver;
+  let resolver: UserResolverQueries;
+  let resolverMutations: UserResolverMutations;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [AppModule],
     }).compile();
 
-    resolver = module.get<UserResolver>(UserResolver);
+    resolver = module.get<UserResolverQueries>(UserResolverQueries);
+    resolverMutations = module.get<UserResolverMutations>(
+      UserResolverMutations
+    );
   });
 
   it('should be defined', () => {
     expect(resolver).toBeDefined();
+    expect(resolverMutations).toBeDefined();
   });
 });
