--- src/utils/aad/aad.account-management.constants.ts
@@ -1,7 +1,0 @@
-export const AAD_ACCOUNT_MANAGEMENT_MODULE_OPTIONS =
-  'aadAccountManagementModuleOptions';
-export const AAD_AUTH_CLIENT_PROVIDER = 'aadAccountManagementModule';
-export const AAD_OBO_PROVIDER = 'aadOboProvider';
-export const AAD_GRAPH_MODULE_PROVIDER = 'aadGraphModule';
-export const AAD_ACCOUNT_MANAGEMENT_MODULE_NEST_PROVIDER =
-  'nestAadAccountManagementModule';
