+++ src/utils/bootstrap/bootstrap.module.ts
@@ -5,12 +5,10 @@ import { EcoverseModule } from '@domain/ecoverse/ecoverse.module';
 import { ProfileModule } from '@domain/profile/profile.module';
 import { TagsetModule } from '@domain/tagset/tagset.module';
 import { UserModule } from '@domain/user/user.module';
-import { AccountModule } from '@utils/account/account.module';
 import { BootstrapService } from './bootstrap.service';
 
 @Module({
   imports: [
-    AccountModule,
     EcoverseModule,
     ProfileModule,
     TagsetModule,
